generalize createController
Move it from EntriesTreeControllerTest to a new ControllerTestBase.
